Include function is missing a close bracket (#376)
The call to include(joinpath("config", "initializers", "searchlight.jl")  is technically missing a close parenthesis. 

This is actually pretty confusing, because the very next line is another statement, but does not begin with the julia REPL prompt, indicating that perhaps this line in intended to be included as a parameter in the previous line's function call? It's unclear. 

Thanks for considering this clarification :)