contains => occursin (needle-haystack around the wrong way) (#577)
`occursin is `occursin(needle, haystack)`, whereas contains is `contains(haystack, needle)`

The logic in Headers was comparing an individual header or origin as the first argument (ie a needle) against the second argument as the haystack (list of origins/allowed headers). So it should use occursin, not contains. (Or flip the order of the arguments, but this way works I think)